feat: make GroupChatRoomCreate customizable
