+++ packages/components/modules/messages/web/GroupChatRoomCreate/Header/styled.tsx
@@ -1,0 +1,11 @@
+import { Box } from '@mui/material'
+import { styled } from '@mui/material/styles'
+
+export const HeaderContainer = styled(Box)(({ theme }) => ({
+  display: 'grid',
+  gridTemplateColumns: '24px auto 24px',
+  alignItems: 'center',
+  width: '100%',
+  padding: theme.spacing(2),
+  borderBottom: `1px solid ${theme.palette.divider}`,
+}))
