+++ packages/components/modules/messages/web/GroupChatRoomCreate/Header/types.ts
@@ -1,0 +1,5 @@
+export interface HeaderProps {
+  isDisabled: boolean
+  onCreateButtonClicked: () => void
+  onBackButtonClicked: () => void
+}
